Moved most classes out of monolithic listbuilder.py; pulling common logic into super class Piece
