+++ listbuilder.py
@@ -1,20 +1,40 @@
 #!/usr/bin/env python3
 
-import shutil
-import os
-import zipfile
+"""
+listbuilder.py
+
+This module provides classes and functions for importing, exporting, and converting Star Wars: Armada fleet lists between various formats and VASSAL-compatible files.
+
+Configuration:
+---------------
+All public functions and classes require a `ShrimpConfig` object, which encapsulates all file paths and runtime options. This eliminates the need for global variables and makes dependencies explicit.
+
+- To use with default settings, create a config with `config = ShrimpConfig()`.
+- To override defaults, pass arguments to `ShrimpConfig(...)`.
+- When calling as a script, command-line arguments are parsed and used to populate a `ShrimpConfig` instance.
+- When importing as a module, always pass a config object to functions/classes.
+
+Example usage (as a module):
+    import listbuilder
+    config = listbuilder.ShrimpConfig(vlb="myfleet.vlb", db="custom_db.vlo")
+    result = listbuilder.import_from_list(config)
+
+"""
+
 import argparse
-import sqlite3
-import re
-import random
 import logging
 import logging.handlers
+import os
+import re
+import shutil
+import sqlite3
 
-from definitions import (
+from listbuilder.definitions import (
     nomenclature_translation,
     ambiguous_names,
 )
-
+from listbuilder.fleet import Fleet
+from listbuilder.utils import unzipall, zipall, scrub_piecename
 
 _handler = logging.handlers.WatchedFileHandler("/var/log/shrimpbot/shrimp.log")
 logging.basicConfig(handlers=[_handler], level=logging.INFO)
@@ -41,29 +61,6 @@ class ShrimpConfig:
         self.fleet = os.path.abspath(flt)  # path or text
         self.db_path = os.path.abspath(db)
         self.import_vlog = import_vlog
-
-
-def unzipall(zip_file_path, tar_path):
-    """Unzips all of the files in the zip file at zip_file_path and
-    dumps all those files into directory tar_path.
-
-    I'm pretty sure this duplicates a built-in function, but, I mean...
-    it works."""
-
-    zip_ref = zipfile.ZipFile(zip_file_path, "r")
-    zip_ref.extractall(tar_path)
-    zip_ref.close()
-
-
-def zipall(src_dir, dest_zip):
-    """Creates a new zip file at zip_file_path and populates it with
-    the zipped contents of tar_path.
-
-    I'm pretty sure this duplicates a built-in function, but, I mean...
-    it works."""
-
-    archive_path = shutil.make_archive(dest_zip, "zip", src_dir)
-    shutil.move(archive_path, dest_zip)
 
 
 def ident_format(fleet_text):
@@ -781,735 +778,8 @@ def export_to_vlog(config):
     zipall(config.working_dir, os.path.abspath(config.vlog_path))
 
 
-def scrub_piecename(piecename):
-    """Scrubs a piece name of most characters that are not alphanumeric; specifically,
-    those that have substantial meaning in SQL queries.  These piecenames are about
-    to be inserted into a SQL query, so we're looking to avoid SQL injection
-    vulnerabilities, as well as just generally wanting clean piecenames."""
-
-    piecename = piecename.replace("\\/", "").split("/")[0].split(";")[-1]
-
-    scrub_these = " :!-'(),\"+.\t\r\n·[]" + "\u2022"
-    for char in scrub_these:
-        piecename = piecename.replace(char, "")
-
-    return piecename.lower()
-
-
 def get_default_config():
     return ShrimpConfig()
-
-
-def calc_guid():
-    return str(round(random.random() * 10**13))
-
-
-class Piece:
-    """Meant to be a prototype for the other pieces, not really to be used on its own"""
-
-    def __init__(self, piecename, config):
-        self.banana = scrub_piecename(str(piecename))
-        self.conn = config.db_path
-        with sqlite3.connect(self.conn) as connection:
-            self.content = connection.execute(
-                """select content from pieces where piecename=?;""", (self.upgradename,)
-            ).fetchall()[0][0]
-
-        self.guid = calc_guid()
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-        self.coords = [0, 0]
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
-
-
-class Fleet:
-    def __init__(
-        self,
-        name,
-        config,
-        faction="",
-        points=0,
-        mode="",
-        fleet_version="",
-        description="",
-        objectives={},
-        ships=[],
-        squadrons=[],
-        author="",
-    ):
-        self.name = str(name)
-        self.faction = str(faction)
-        self.points = int(points)
-        self.mode = str(mode)
-        self.fleet_version = str(fleet_version)
-        self.description = str(description)
-        self.objectives = dict(objectives)
-        self.ships = list(ships)
-        self.squadrons = list(squadrons)
-        self.author = str(author)
-        self.conn = config.db_path
-
-        # simple piece locations calculations
-
-        self.x = 200
-        self.ship_y = 850
-        self.upgd_upper_y = 775
-
-        self.sc_to_obj_x_padding = -100
-        self.sc_to_obj_y_padding = 0
-        self.obj_x = self.x + self.sc_to_obj_x_padding
-        self.obj_y = self.ship_y + self.sc_to_obj_y_padding
-
-        self.cmd_to_sc_x_offset = 90
-        self.cmd_to_sc_y_offset = -10
-
-        self.sc_to_st_x_padding = 173
-        self.sc_to_st_y_padding = 0
-        # self.s_to_u_padding = 60
-        # self.u_to_u_x_padding = 105
-        # self.u_to_u_y_padding = 175
-        self.s_to_u_padding = 50
-        self.u_to_u_x_padding = 145
-        self.u_to_u_y_padding = 225
-        self.u_to_s_padding = 195
-        self.u_to_sq_padding = 195
-        self.upgd_lower_y = self.upgd_upper_y + self.u_to_u_y_padding
-        self.sq_to_sq_x_padding = 175
-        self.sq_to_sq_y_padding = 240
-        self.obj_to_obj_x_offset = 25
-        self.obj_to_obj_y_offset = 25
-
-        self.sq_y_offset = -120
-        self.sq_upper_y = self.ship_y + self.sq_y_offset
-        self.sq_lower_y = self.sq_upper_y + self.sq_to_sq_y_padding
-        self.sq_row = 1
-
-    def set_name(self, name):
-        self.name = str(name)
-
-    def set_faction(self, faction):
-        self.faction = str(faction)
-
-    def set_points(self, points):
-        if points.isdigit():
-            self.points = int(points)
-            return
-        logging.info("Failed to set points: '{}' cannot be converted to an integer.")
-        self.points = int(0)
-        return
-
-    def set_mode(self, mode):
-        self.mode = str(mode)
-
-    def set_fleet_version(self, fleet_version):
-        self.fleet_version = str(fleet_version)
-
-    def set_description(self, description):
-        self.description = str(description)
-
-    def set_objectives(self, objectives):
-        self.objectives = dict(objectives)
-
-    def add_ship(self, shipclass):
-        shipclass = scrub_piecename(shipclass)
-        if shipclass in nomenclature_translation:
-            sc = nomenclature_translation[shipclass]
-            logging.info(
-                "[-] Translated {} to {} - in listbuilder.Fleet.add_ship().".format(
-                    shipclass, sc
-                )
-            )
-            shipclass = sc
-
-        s = Ship(shipclass, self, self.conn)
-        self.x += self.u_to_s_padding
-        s.set_coords([str(self.x), str(self.ship_y)])
-        s.shipcard.set_coords([str(self.x), str(self.ship_y)])
-        s.shipcmdstack.set_coords(
-            [
-                str(self.x + self.cmd_to_sc_x_offset),
-                str(self.ship_y + self.cmd_to_sc_y_offset),
-            ]
-        )
-        self.x += self.sc_to_st_x_padding
-        s.shiptoken.set_coords(
-            [str(self.x), str(self.ship_y + self.sc_to_st_y_padding)]
-        )
-        self.x += self.s_to_u_padding
-        self.u_row = 1
-
-        self.ships.append(s)
-        return s
-
-    def remove_ship(self, ship):
-        self.ships.remove(ship)
-
-    def add_squadron(self, squadronclass):
-        squadronclass = scrub_piecename(squadronclass)
-        if squadronclass in nomenclature_translation:
-            sc = nomenclature_translation[squadronclass]
-            logging.info(
-                "[-] Translated {} to {} - in listbuilder.Fleet.add_squadron().".format(
-                    squadronclass, sc
-                )
-            )
-            squadronclass = sc
-
-        try:
-            sq = Squadron(squadronclass, self, self.conn)
-        except ValueError as err:
-            """If the squadron is not found in the database, try scrubbing the word "squadron" and retrying."""
-            squadronclass = scrub_piecename(squadronclass.replace("squadron", ""))
-            sq = Squadron(squadronclass, self, self.conn)
-        if self.sq_row % 2:
-            self.x += self.sq_to_sq_x_padding
-            sq.set_coords([str(self.x), str(self.sq_upper_y)])
-            sq.squadroncard.set_coords([str(self.x), str(self.sq_upper_y)])
-            sq.squadrontoken.set_coords([str(self.x), str(self.sq_upper_y)])
-        else:
-            sq.set_coords([str(self.x), str(self.sq_lower_y)])
-            sq.squadroncard.set_coords([str(self.x), str(self.sq_lower_y)])
-            sq.squadrontoken.set_coords([str(self.x), str(self.sq_lower_y)])
-        self.sq_row += 1
-
-        self.squadrons.append(sq)
-        return sq
-
-    def remove_squadron(self, squadron):
-        self.squadrons.remove(squadron)
-
-    def add_objective(self, category, objectivename):
-        category = scrub_piecename(category)
-        objectivename = scrub_piecename(objectivename)
-
-        if objectivename in nomenclature_translation:
-            ob = nomenclature_translation[objectivename]
-            logging.info(
-                "[-] Translated {} to {} - in listbuilder.Fleet.add_objective()".format(
-                    objectivename, ob
-                )
-            )
-            objectivename = ob
-
-        if "custom" in objectivename.lower():
-            return False
-
-        obj_categories = ["assault", "defense", "navigation", "campaign", "other"]
-        if category.lower() in obj_categories:
-            self.objectives[category] = Objective(objectivename, self.conn)
-        else:
-            # except:
-            # raise ValueError
-            logging.info("{} is not a valid objective type.".format(str(category)))
-            logging.info("Valid types are: {}".format(obj_categories))
-
-        self.objectives[category].set_coords([str(self.obj_x), str(self.obj_y)])
-        self.obj_x = self.obj_x + self.obj_to_obj_x_offset
-        self.obj_y = self.obj_y + self.obj_to_obj_y_offset
-
-    def remove_objective(self, category, objective):
-        if category in self.objectives.keys():
-            if self.objectives[category] == objective:
-                del self.objectives[category]
-
-    def __add__(self, ship):
-        self.add_ship(ship)
-
-    def __sub__(self, ship):
-        self.remove_ship(ship)
-
-
-class Ship:
-    def __init__(self, shipclass, ownfleet, config):
-        self.shipclass = scrub_piecename(str(shipclass))  # "name" in .AFF
-        self.conn = config.db_path
-        self.content = ""
-        self.coords = [0, 0]
-        self.physicalsize = [
-            [0, 0],
-            [0, 0],
-        ]  # amt of table space for shipcard, stack, and all upgrades
-        self.shipcard = ShipCard(self.shipclass, self.conn)
-        self.shiptoken = self.shipcard.shiptoken
-        self.shipcmdstack = self.shipcard.shipcmdstack
-        self.upgrades = []
-        self.guid = calc_guid()
-
-        self.ownfleet = ownfleet
-
-    def set_content(self, content):
-        self.content = str(content)
-
-    def set_coords(self, coords):
-        self.coords = list(coords)
-
-    def set_shipcard(self, shipcard):
-        self.shipcard = shipcard
-
-    def set_shiptoken(self, shiptoken):
-        self.shiptoken = shiptoken
-
-    def set_upgrades(self, upgrades):
-        self.upgrades = list(upgrades)
-
-    def add_upgrade(self, upgradename):
-        upgradename = scrub_piecename(upgradename)
-        if upgradename in nomenclature_translation:
-            sc = nomenclature_translation[upgradename]
-            logging.info(
-                "[-] Translated {} to {} - in listbuilder.Ship.add_upgrade().".format(
-                    upgradename, sc
-                )
-            )
-            upgradename = sc
-
-        u = Upgrade(upgradename, self, config)
-
-        if self.ownfleet.u_row % 2:
-            self.ownfleet.x += self.ownfleet.u_to_u_x_padding
-            u.set_coords([str(self.ownfleet.x), str(self.ownfleet.upgd_upper_y)])
-        else:
-            u.set_coords([str(self.ownfleet.x), str(self.ownfleet.upgd_lower_y)])
-        self.ownfleet.u_row += 1
-
-        self.upgrades.append(u)
-        return u
-
-    def remove_upgrade(self, upgrade):
-        self.upgrades.remove(upgrade)
-
-    def __add__(self, upgrade):
-        self.add_upgrade(upgrade)
-
-    def __sub__(self, upgrade):
-        self.remove_upgrade(upgrade)
-
-
-class ShipCard:
-    """A shipcard of type str(shipname) as defined in sqlitedb connection conn."""
-
-    def __init__(self, shipname, config):
-        self.shipname = scrub_piecename(str(shipname))
-        self.conn = config.db_path
-
-        logging.info(
-            "Searching for ship {} in {}".format(self.shipname, str(self.conn))
-        )
-
-        try:
-            with sqlite3.connect(self.conn) as connection:
-                logging.debug("0")
-                exact_match = connection.execute(
-                    """select content,catchall 
-                    from pieces 
-                    where piecetype='shipcard' 
-                    and piecename=?;""",
-                    (self.shipname,),
-                ).fetchall()
-                logging.debug(".5")
-            logging.debug("1")
-            # logging.info(str(exact_match))
-            # logging.info(str(len(exact_match)))
-            if len(exact_match) == 1:
-                logging.debug("2")
-                [(self.content, self.shiptype)] = exact_match
-                logging.debug("3")
-            if not (hasattr(self, "content") and hasattr(self, "shiptype")):
-                logging.debug("4")
-                raise RuntimeError(f"Did not find ship card {self.shipname}")
-        except RuntimeError as err:
-            logging.debug("5")
-            logging.debug(err, exc_info=err)
-            logging.debug("6")
-        except Exception as err:
-            logging.debug("7")
-            logging.debug(err, exc_info=err)
-            raise err
-
-        logging.debug("8")
-        self.shiptoken = ShipToken(self.shiptype, self.conn)
-
-        self.guid = calc_guid()
-        logging.debug("9")
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-
-        logging.debug("10")
-
-        self.command = self.content.split("/placemark;Spawn Command ")[-1][0]
-        self.command = "commandstack" + self.command
-        self.shipcmdstack = ShipCmdStack(self.command, self.conn)
-
-        self.coords = [0, 0]
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
-
-    def set_guid(self, guid):
-        self.content = self.content.replace("vlb_GUID", self.guid)
-
-    def set_shiptoken(self, shiptype):
-        self.shiptoken = ShipToken(shiptype, self.conn)
-
-
-class ShipToken:
-    def __init__(self, shiptype, config):
-        self.shiptype = scrub_piecename(str(shiptype))
-        self.conn = config.db_path
-
-        if self.shiptype in nomenclature_translation:
-            translated_shiptype = nomenclature_translation[self.shiptype]
-            logging.info(
-                "[-] Translated {} to {} - in listbuilder.ShipToken.".format(
-                    self.shiptype, translated_shiptype
-                )
-            )
-            self.shiptype = translated_shiptype
-
-        logging.info(
-            "Searching for ship token {} in {}".format(self.shiptype, str(self.conn))
-        )
-
-        try:
-            with sqlite3.connect(self.conn) as connection:
-                exact_match = connection.execute(
-                    """select content 
-                    from pieces 
-                    where piecetype='ship' and piecename=?;""",
-                    (self.shiptype,),
-                ).fetchall()
-            if len(exact_match) == 1:
-                if len(exact_match[0]) == 1:
-                    self.content = exact_match[0][0]
-            if not self.content:
-                raise RuntimeError(f"Did not find ship token {self.shiptype}")
-        except RuntimeError as err:
-            logging.exception(err)
-            raise err
-        except Exception as err:
-            logging.debug(err, exc_info=err)
-            raise err
-
-        self.guid = calc_guid()
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-        self.coords = [0, 0]
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
-
-
-class ShipCmdStack:
-    """A command stack as defined in sqlitedb connection conn."""
-
-    def __init__(self, cmdstack, config):
-        self.cmdstack = scrub_piecename(str(cmdstack))
-        self.conn = config.db_path
-
-        logging.info(
-            "Searching for command stack {} in {}".format(self.cmdstack, str(self.conn))
-        )
-        try:
-            with sqlite3.connect(self.conn) as connection:
-                exact_match = connection.execute(
-                    """select content 
-                    from pieces 
-                    where piecetype='other' and piecename=?;""",
-                    (self.cmdstack,),
-                ).fetchall()
-            if len(exact_match) == 1:
-                if len(exact_match[0]) == 1:
-                    self.content = exact_match[0][0]
-            if not self.content:
-                raise RuntimeError(f"Did not find command stack {self.cmdstack}")
-        except RuntimeError as err:
-            logging.exception(err)
-            raise err
-        except Exception as err:
-            logging.debug(err, exc_info=err)
-            raise err
-
-        self.guid = calc_guid()
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-        self.coords = [0, 0]
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
-
-    def set_guid(self, guid):
-        self.content = self.content.replace("vlb_GUID", self.guid)
-
-
-class Upgrade:
-    def __init__(self, upgradename, ownship, config):
-        self.upgradename = scrub_piecename(str(upgradename))
-        self.conn = config.db_path
-
-        logging.info(
-            "Searching for upgrade {} in {}".format(self.upgradename, str(self.conn))
-        )
-
-        try:
-            with sqlite3.connect(self.conn) as connection:
-                exact_match = connection.execute(
-                    """select content 
-                    from pieces 
-                    where piecetype='upgradecard' and piecename=?;""",
-                    (self.upgradename,),
-                ).fetchall()
-
-            self.content = False
-            if len(exact_match) == 1:
-                if len(exact_match[0]) == 1:
-                    self.content = exact_match[0][0]
-            if not self.content:
-                raise RuntimeError(f"Did not find upgrade {self.upgradename}")
-        except RuntimeError as err:
-            logging.exception(err)
-            raise err
-        except Exception as err:
-            logging.exception(err)
-            raise err
-
-        self.guid = calc_guid()
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-        self.coords = [0, 0]
-
-        self.ownship = ownship
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
-
-
-class Squadron:
-    def __init__(self, squadronclass, ownfleet, config):
-        self.squadronclass = scrub_piecename(str(squadronclass))  # "name" in .AFF
-        self.conn = config.db_path
-        self.content = ""
-        self.coords = [0, 0]
-        self.squadroncard = SquadronCard(self.squadronclass, self.conn)
-        self.squadrontoken = self.squadroncard.squadrontoken
-        self.upgrades = []
-        self.guid = calc_guid()
-
-        self.ownfleet = ownfleet
-
-    def set_content(self, content):
-        self.content = str(content)
-
-    def set_coords(self, coords):
-        self.coords = list(coords)
-
-    def set_squadroncard(self, squadroncard):
-        self.squadroncard = squadroncard
-
-    def set_squadrontoken(self, squadrontoken):
-        self.squadrontoken = squadrontoken
-
-
-class SquadronCard:
-    """A squadroncard of type str(squadronname) as defined in sqlite connection conn."""
-
-    def __init__(self, squadronname, config):
-        self.squadronname = scrub_piecename(str(squadronname))
-        self.conn = config.db_path
-
-        logging.info(
-            "Searching for squadron card {} in {}".format(
-                self.squadronname, str(self.conn)
-            )
-        )
-
-        try:
-            with sqlite3.connect(self.conn) as connection:
-                exact_match = connection.execute(
-                    """select content,catchall 
-                    from pieces 
-                    where piecetype='squadroncard' and piecename like ?;""",
-                    (self.squadronname,),
-                ).fetchall()
-            if len(exact_match) == 1:
-                [(self.content, self.squadrontype)] = exact_match
-            else:
-                with sqlite3.connect(self.conn) as connection:
-                    [(self.content, self.squadrontype)] = connection.execute(
-                        """select content,catchall 
-                        from pieces 
-                        where piecetype='squadroncard' and piecename like ?;""",
-                        ("%" + self.squadronname + "%",),
-                    ).fetchall()
-        except ValueError as err:
-            logging.exception(f"Did not find squadron {self.squadronname}")
-            raise err
-        except Exception as err:
-            logging.debug(err, exc_info=err)
-            raise err
-
-        self.squadrontoken = SquadronToken(self.squadrontype, self.conn)
-
-        # self.guid = str(round(time.time()*1000))
-        self.guid = calc_guid()
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-        self.coords = [0, 0]
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
-
-    def set_guid(self, guid):
-        self.content = self.content.replace("vlb_GUID", self.guid)
-
-    def set_squadrontoken(self, squadrontype):
-        self.squadrontoken = SquadronToken(squadrontype, self.conn)
-
-
-class SquadronToken:
-    def __init__(self, squadrontype, config):
-        self.squadrontype = scrub_piecename(str(squadrontype))
-        self.conn = config.db_path
-        logging.info(
-            "Searching for squadron token {} in {}".format(
-                scrub_piecename(squadrontype), str(conn)
-            )
-        )
-
-        try:
-            with sqlite3.connect(self.conn) as connection:
-                exact_match = connection.execute(
-                    """select content 
-                    from pieces 
-                    where piecetype='squadron' and piecename=?;""",
-                    (self.squadrontype,),
-                ).fetchall()
-            if len(exact_match) == 1:
-                if len(exact_match[0]) == 1:
-                    self.content = exact_match[0][0]
-            if not self.content:
-                raise RuntimeError(f"Did not find squadron token {self.squadrontype}")
-        except RuntimeError as err:
-            logging.exception(err)
-            raise err
-        except Exception as err:
-            logging.debug(err, exc_info=err)
-            raise err
-
-        self.guid = calc_guid()
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-        self.coords = [0, 0]
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
-
-
-class Objective:
-    def __init__(self, objectivename, config):
-        self.objectivename = scrub_piecename(str(objectivename))
-        self.conn = config.db_path
-        logging.info(
-            "Searching for objective {} in {}".format(
-                self.objectivename, str(self.conn)
-            )
-        )
-
-        try:
-            with sqlite3.connect(self.conn) as connection:
-                exact_match = connection.execute(
-                    """select content 
-                    from pieces 
-                    where piecetype='objective' and piecename=?;""",
-                    (self.objectivename,),
-                ).fetchall()
-            if len(exact_match) == 1:
-                if len(exact_match[0]) == 1:
-                    self.content = exact_match[0][0]
-            if not self.content:
-                raise RuntimeError(f"Did not find objective {self.objectivename}")
-        except RuntimeError as err:
-            logging.exception(err)
-            raise err
-        except Exception as err:
-            logging.debug(err, exc_info=err)
-            raise err
-
-        self.guid = calc_guid()
-        self.content = self.content.replace("vlb_GUID", self.guid)
-        self.content = self.content.replace("vlb_x_axis", "0")
-        self.content = self.content.replace("vlb_y_axis", "0")
-
-        c = ""
-        for line in self.content.split("\t"):
-            if line.strip().startswith("piece;;;;"):
-                this_line = line.replace("1", "2")
-            else:
-                this_line = line
-            c += this_line + "\t"
-
-        self.content = c
-
-        self.coords = [0, 0]
-
-    def set_coords(self, coords):
-        if type(coords) == list and len(coords) == 2:
-            self.content = re.sub(
-                r"Table;\d{1,4};\d{1,4}",
-                "Table;{};{}".format(str(coords[0]), str(coords[1])),
-                self.content,
-            )
-            self.coords = coords
 
 
 if __name__ == "__main__":
